# app/cmd/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "RSS feed aggregator with passkey authentication.",
        "title": "GKFeed API",
        "contact": {
            "name": "GKFeed",
            "url": "https://github.com/gkfeed/api-go"
        },
        "version": "2.0"
    },
    "paths": {
...
        "/api/v1/delete": {
            "delete": {
                "security": [
                    {
                        "BasicAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Deletes a feed by ID. Only the feed owner can delete it.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "feeds"
                ],
                "summary": "Delete feed",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Feed ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": ""
                    },
                    "400": {
                        "description": ""
                    },
                    "401": {
                        "description": ""
                    },
                    "404": {
                        "description": ""
                    },
                    "500": {
                        "description": ""
                    }
                }
            }
        },
...
        "/api/v1/items/{id}": {
            "delete": {
                "security": [
                    {
                        "BasicAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Permanently deletes an item owned by the authenticated user.",
                "tags": [
                    "items"
                ],
                "summary": "Delete item",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Item ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": ""
                    },
                    "400": {
                        "description": ""
                    },
                    "401": {
                        "description": ""
                    },
                    "404": {
                        "description": ""
                    },
                    "500": {
                        "description": ""
                    }
                }
            }
        },
...
        "/api/v2/feeds/{id}": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns a feed owned by the authenticated user.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "feeds"
                ],
                "summary": "Get feed by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Feed ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.feedDTO"
                        }
                    },
                    "400": {
                        "description": ""
                    },
                    "401": {
                        "description": ""
                    },
                    "404": {
                        "description": ""
                    },
                    "500": {
                        "description": ""
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BasicAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Deletes a feed and all its items. Only the owner can delete it.",
                "tags": [
                    "feeds"
                ],
                "summary": "Delete feed by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Feed ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": ""
                    },
                    "400": {
                        "description": ""
                    },
                    "401": {
                        "description": ""
                    },
                    "404": {
                        "description": ""
                    },
                    "500": {
                        "description": ""
                    }
                }
            }
        },
...
        "/api/v2/items/changes": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns item mutations after a sync cursor. Apply each response before using its next cursor.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "items"
                ],
                "summary": "Get item changes",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Changes per page (default 100, maximum 500)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Opaque sync cursor",
                        "name": "cursor",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.itemChangesResponse"
                        }
                    },
                    "400": {
                        "description": ""
                    },
                    "401": {
                        "description": ""
                    },
                    "500": {
                        "description": ""
                    }
                }
            }
        },
        "/api/v2/items/sync": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns a bounded item page and a sync cursor for later changes.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "items"
                ],
                "summary": "List items for synchronization",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Items per page (default 100, maximum 500)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Opaque page cursor",
                        "name": "cursor",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.syncItemsResponse"
                        }
                    },
                    "400": {
                        "description": ""
                    },
                    "401": {
                        "description": ""
                    },
                    "500": {
                        "description": ""
                    }
                }
            }
        },
        "/api/v2/items/{id}": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns an item and its parent feed for the authenticated user.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "items"
                ],
                "summary": "Get item by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Item ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.getItemResponse"
                        }
                    },
                    "400": {
                        "description": ""
                    },
                    "401": {
                        "description": ""
                    },
                    "404": {
                        "description": ""
                    },
                    "500": {
                        "description": ""
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BasicAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Permanently deletes an item owned by the authenticated user.",
                "tags": [
                    "items"
                ],
                "summary": "Delete item",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Item ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": ""
                    },
                    "400": {
                        "description": ""
                    },
                    "401": {
                        "description": ""
                    },
                    "404": {
                        "description": ""
                    },
                    "500": {
                        "description": ""
                    }
                }
            }
        }
    },
    "definitions": {
...
        "handlers.feedDTO": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                },
                "userid": {
                    "type": "integer"
                }
            }
        },
...
        "handlers.getItemResponse": {
            "type": "object",
            "properties": {
                "feed": {
                    "$ref": "#/definitions/handlers.feedDTO"
                },
                "item": {
                    "$ref": "#/definitions/handlers.itemDTO"
                }
            }
        },
...
        "handlers.itemChangesResponse": {
            "type": "object",
            "properties": {
                "deleted_ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "has_more": {
                    "type": "boolean"
                },
                "next_cursor": {
                    "type": "string"
                },
                "upserted": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handlers.itemDTO"
                    }
                }
            }
        },
        "handlers.itemDTO": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string"
                },
                "feed_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "link": {
                    "type": "string"
                },
                "text": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "handlers.syncItemsResponse": {
            "type": "object",
            "properties": {
                "has_more": {
                    "type": "boolean"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handlers.itemDTO"
                    }
                },
                "next_cursor": {
                    "type": "string"
                },
                "sync_cursor": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        },
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}